More changes for temporary tables: - Add a catalog entity to back temporary_tables and global_temporary_tables.  Add   info schema tables for same. - Arrange for the emitter to emit the TEMPORARY keyword appropriately. - Add a new schema subtype (TemporaryTableSchema) to hold temporary table definitions with their   interesting lookup rules. - Switch DDLCallback and NestedDDLCallback from interfaces to abstract classes.   Add two new methods on DDLCallback: onRollback & onCommit.  Use these to arrange   to populate the TemporaryTableSchema on the context for the connection as appropriate   on create table.  Still to do: use them to handle drop table. - Introduce a new class, a TableResolver, which encapsulates the resolution logic.   Use it in place of bespoke methods.  The primary purpose of this class is to unify   table resolution code, and make it easier to insert the temporary table lookup.   There are still some spots using the existing bespoke methods, these need to be   cleaned up. - More changes in ComplexPETable to handle temporary tables - in particular ensure   they have no persistent representation. - Add worker group pinning.  Worker Groups keep a counter of how many temp tables are   on them, and will not allow themselves to be purged if any temp tables are outstanding.

+++ tesora-dve-core/src/main/java/com/tesora/dve/server/connectionmanager/SSConnection.java
@@ -850,14 +850,17 @@ public class SSConnection extends Agent implements WorkerGroup.Manager, LockClie
 	@Override
 	public void returnWorkerGroup(WorkerGroup wg) throws PEException {
 		StorageGroup sg = wg.getGroup();
-		wg.disassociateFromConnection();
+		if (!wg.isPinned())
+			wg.disassociateFromConnection();
 		if (activeWG.containsKey(sg) && wg == activeWG.get(sg)) {
 			activeWG.remove(sg);
-			if (wg.isMarkedForPurge())
+			if (wg.isMarkedForPurge() && !wg.isPinned())
 				WorkerGroupFactory.purgeInstance(this, wg);
 			else
 				availableWG.put(sg, wg);
 		} else {
+			if (wg.isPinned())
+				return;
 			// This snippet addresses an issue where worker groups are given 
 			// back to the wg factory, even though they are still in the available list
 			boolean stillInAvailableList = false;
